test-avocado: Change it completely.
