--- test-avocado/lib/phantom-driver
@@ -1,1 +1,0 @@
-../../test/phantom-driver
