--- test-avocado/lib/phantom-lib.js
@@ -1,1 +1,0 @@
-../../test/phantom-lib.js
