--- test-avocado/lib/sizzle.js
@@ -1,1 +1,0 @@
-../../lib/sizzle.v2.1.0.js
