--- test-avocado/lib/testlib.py
@@ -1,1 +1,0 @@
-../../test/testlib.py
