--- test-avocado/lib/testvm.py
@@ -1,1 +1,0 @@
-../../test/testvm.py
