+++ test-avocado/phantom-lib.js
@@ -1,0 +1,1 @@
+../test/phantom-lib.js
