+++ test-avocado/sizzle.js
@@ -1,0 +1,1 @@
+../test/sizzle.js
