+++ test-avocado/testlib.py
@@ -1,0 +1,1 @@
+../test/testlib.py
