+++ test-avocado/testvm.py
@@ -1,0 +1,1 @@
+../test/testvm.py
