+++ test/README
@@ -3,6 +3,13 @@
 This directory contains automated tests for Cockpit, and the support
 files for them.
 
+  **********************************************************************
+  *                                                                    *
+  *  This testsuite is being phased out.  Don't do any new work here.  *
+  *  The new one is in ../test-avocado.                                *
+  *                                                                    *
+  **********************************************************************
+
 To run the tests, you need to install the following packages on
 Fedora:
